feat: add initial web implementation
